Completed more game logic, made io generic for running wumpus in both cosole and web mode

+++ src/actor/actor.rs
@@ -9,5 +9,5 @@ pub enum ActorType {
 #[derive(Copy, Clone)]
 pub struct Actor {
     pub actor_type: ActorType,
-    pub room: u16,
+    pub room: u8,
 }
